feat: rewrite v4 native-stack to TS
