+++ createNativeStackNavigator/package.json
@@ -1,5 +1,6 @@
 {
   "main": "../lib/commonjs/createNativeStackNavigator",
   "module": "../lib/module/createNativeStackNavigator",
-  "react-native": "../src/createNativeStackNavigator"
+  "react-native": "../src/createNativeStackNavigator",
+  "types": "../lib/typescript/createNativeStackNavigator.d.ts"
 }
